+++ expansion_types.c
@@ -1,0 +1,33 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   expansion_types.c                                  :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: marcudos <marcudos@student.42sp.org.br>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/04/28 16:57:12 by marcudos          #+#    #+#             */
+/*   Updated: 2025/04/28 16:57:36 by marcudos         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+char	*expand_env(char *input, t_env *env)
+{
+	char	*key;
+
+	key = extract_key(input);
+	if (!key)
+		return (ft_strdup(""));
+	while (env)
+	{
+		if (ft_strcmp(env->key, key) == 0)
+		{
+			free (key);
+			return (ft_strdup(env->value));
+		}
+		env = env->next;
+	}
+	free(key);
+	return (ft_strdup(""));
+}
